+++ docs/guides/changelog.md
@@ -16,6 +16,7 @@ description: |-
 - Add `sdwan_topology_custom_control_feature` resource and data source
 - Add `sdwan_topology_mesh_feature` resource and data source
 - Add `sdwan_topology_hub_spoke_feature` resource and data source
+- Add `sdwan_custom_application` resource and data source
 
 ## 0.11.2